refactor: Separate Domain and Data layers for Entities/DTOs
- Create lib/src/data layer for DTOs (Models).
- Move all *_model.dart files from domain to data layer.
- Refactor Models to remove inheritance from Entities.
- Implement toEntity/fromEntity mapping methods in Models.
- Add ARCHITECTURE_REFACTOR.md documentation.
