+++ lib/.dart_tool/package_config_subset
@@ -1,0 +1,41 @@
+characters
+3.4
+file:///Users/adiala/.pub-cache/hosted/pub.dev/characters-1.4.0/
+file:///Users/adiala/.pub-cache/hosted/pub.dev/characters-1.4.0/lib/
+collection
+3.4
+file:///Users/adiala/.pub-cache/hosted/pub.dev/collection-1.19.1/
+file:///Users/adiala/.pub-cache/hosted/pub.dev/collection-1.19.1/lib/
+flutter_lints
+2.19
+file:///Users/adiala/.pub-cache/hosted/pub.dev/flutter_lints-2.0.3/
+file:///Users/adiala/.pub-cache/hosted/pub.dev/flutter_lints-2.0.3/lib/
+lints
+3.0
+file:///Users/adiala/.pub-cache/hosted/pub.dev/lints-2.1.1/
+file:///Users/adiala/.pub-cache/hosted/pub.dev/lints-2.1.1/lib/
+material_color_utilities
+2.17
+file:///Users/adiala/.pub-cache/hosted/pub.dev/material_color_utilities-0.11.1/
+file:///Users/adiala/.pub-cache/hosted/pub.dev/material_color_utilities-0.11.1/lib/
+meta
+2.12
+file:///Users/adiala/.pub-cache/hosted/pub.dev/meta-1.16.0/
+file:///Users/adiala/.pub-cache/hosted/pub.dev/meta-1.16.0/lib/
+vector_math
+2.14
+file:///Users/adiala/.pub-cache/hosted/pub.dev/vector_math-2.1.4/
+file:///Users/adiala/.pub-cache/hosted/pub.dev/vector_math-2.1.4/lib/
+sky_engine
+3.7
+file:///Users/adiala/Core/flutter/bin/cache/pkg/sky_engine/
+file:///Users/adiala/Core/flutter/bin/cache/pkg/sky_engine/lib/
+flutter
+3.7
+file:///Users/adiala/Core/flutter/packages/flutter/
+file:///Users/adiala/Core/flutter/packages/flutter/lib/
+tatoh_domain_entity
+3.1
+file:///Users/adiala/Projects/Tatoh/mobile/tatoh_domain_entity/lib/
+file:///Users/adiala/Projects/Tatoh/mobile/tatoh_domain_entity/lib/lib/
+2
