+++ lib/pubspec.yaml
@@ -1,0 +1,15 @@
+name: tatoh_domain_entity
+description: Domain entities for the Tatoh project.
+publish_to: 'none' # Prevents accidental publishing
+
+environment:
+  sdk: '>=3.1.3 <4.0.0' # Match the main project's SDK constraint
+
+dependencies:
+  flutter:
+    sdk: flutter
+  # Add other dependencies if this package needs them (e.g., equatable)
+  # equatable: ^2.0.5
+
+dev_dependencies:
+  flutter_lints: ^2.0.0
