--- lib/src/domain/user/models/user_model.dart
@@ -1,62 +1,0 @@
-import '../entities/user_entity.dart';
-
-/// Modelo para serialização/deserialização de usuários
-class UserModel extends UserEntity {
-  /// Construtor
-  UserModel({
-    required String id,
-    required String email,
-    required String name,
-    String? photoUrl,
-    required bool emailVerified,
-    required DateTime createdAt,
-    required DateTime updatedAt,
-  }) : super(
-          id: id,
-          email: email,
-          name: name,
-          photoUrl: photoUrl,
-          emailVerified: emailVerified,
-          createdAt: createdAt,
-          updatedAt: updatedAt,
-        );
-  
-  /// Cria um modelo a partir de um Map (JSON)
-  factory UserModel.fromJson(Map<String, dynamic> json) {
-    return UserModel(
-      id: json['id'] as String,
-      email: json['email'] as String,
-      name: json['name'] as String,
-      photoUrl: json['photo_url'] as String?,
-      emailVerified: json['email_verified'] as bool,
-      createdAt: DateTime.parse(json['created_at'] as String),
-      updatedAt: DateTime.parse(json['updated_at'] as String),
-    );
-  }
-  
-  /// Converte o modelo para um Map (JSON)
-  Map<String, dynamic> toJson() {
-    return {
-      'id': id,
-      'email': email,
-      'name': name,
-      'photo_url': photoUrl,
-      'email_verified': emailVerified,
-      'created_at': createdAt.toIso8601String(),
-      'updated_at': updatedAt.toIso8601String(),
-    };
-  }
-  
-  /// Cria um modelo a partir de uma entidade
-  factory UserModel.fromEntity(UserEntity entity) {
-    return UserModel(
-      id: entity.id,
-      email: entity.email,
-      name: entity.name,
-      photoUrl: entity.photoUrl,
-      emailVerified: entity.emailVerified,
-      createdAt: entity.createdAt,
-      updatedAt: entity.updatedAt,
-    );
-  }
-}
